+++ frontend_second/src/pages/index.tsx
@@ -10,6 +10,9 @@ export default function Home() {
         <h1>アウトドア般若心経</h1>
         <p>自分探しならぬ、自分なくしの旅へ</p>
       </div>
+      <div>
+        <a href='/welcome'>welcomeページへ</a>
+      </div>
     </>
   )
 }
